requirements.txt
groq
google-genai
pymupdf
chromadb
streamlit
python-dotenv
pillow
pysqlite3-binary
